requirements.txt
einops>=0.3.2
h5py>=2.10.0
mmcv>=1.3.15
numpy>=1.18.5
opencv-python>=4.5.3.56
scipy>=1.5.2
matplotlib>=3.3.1
pillow>=7.2.0
pandas>=1.1.3
torchvision>=0.8.0